==> _analysis/STEP2_farm_and_feed_raster_prep.Rmd <==
---
title: "Summarizing feed and farm pressures"
output: html_document
editor_options: 
  chunk_output_type: console
---

Organizing files and summarizing feed files.

STEP 1: Organize farm and wildcaught files:
new home: /home/shares/food-systems/Food_footprint/all_food_systems/analysis/raw
1. move the livestock and aquaculture farm rasters with prefix to filename "farm_"
2. move the marine and freshwater fisheries rasters with prefix to filename "wildcaught_"

STEP 2: Create system x feed stressor combos
1. multiple _x_ times appropriate pressures and move with prefix to filename "feed_":
/home/shares/food-systems/Food_footprint/all_food_systems/analysis/raw_feed_combos
2. summarize for each food system and stressor feed costs for fofm, fodder, and crops

First part takes about 3 hours to run.
Second part takes about 1 hour to run.

```{r}
library(rgdal)
library(sp)
library(raster)
library(tidyverse)
library(here)
library(foreach)
library(doParallel)

source(here("_workflow/common.R"))
source(here("_workflow/common_spatial.R"))

```

# clear out listed files in tif folders. Want to make sure we have a clean start.
```{r}

do.call(file.remove, list(list.files("/home/shares/food-systems/Food_footprint/all_food_systems/analysis/raw", full.names = TRUE)))

do.call(file.remove, list(list.files("/home/shares/food-systems/Food_footprint/all_food_systems/analysis/raw_feed_combos", full.names = TRUE)))

```


```{r}

complete_list <- read_csv(here("_analysis/food_system_categories.csv"))

# livestock and aquaculture
farm_animal_list <- filter(complete_list, subcategory %in% c("livestock", "aquaculture")) %>%
  mutate(farm = paste(origin, organism, system, product, sep="_")) %>%
  select(farm) %>%
  pull(.)

## N = 27 * 4 = 108
farm_animals <- list.files('/home/shares/food-systems/Food_footprint/all_food_systems/datalayers', pattern = ".tif", full=TRUE)
farm_animals <- grep(paste(farm_animal_list, collapse="|"), farm_animals, value=TRUE)
farm_animals <- grep("_x_", farm_animals, value=TRUE, invert=TRUE)

# copy these files over:
file.copy(farm_animals, "/home/shares/food-systems/Food_footprint/all_food_systems/analysis/raw")
add_farm <- paste0("/home/shares/food-systems/Food_footprint/all_food_systems/analysis/raw/",basename(farm_animals))
file.rename(from= add_farm, 
            to = gsub("/raw/", "/raw/farm_", add_farm))


## Wild caught fisheries
fisheries_list <- filter(complete_list, subcategory %in% c("fisheries")) %>%
  mutate(system = paste(origin, organism, system, product, sep="_")) %>%
  select(system) %>%
  filter(system != "marine_fofm_fisheries_meat") %>%   # this one is accounted for mostly in feed
  pull(.)

# N = 7 * 4 = 28
fisheries <- list.files('/home/shares/food-systems/Food_footprint/all_food_systems/datalayers', pattern = ".tif", full=TRUE)
fisheries <- grep(paste(fisheries_list, collapse="|"), fisheries, value=TRUE)
fisheries <- grep("_x_", fisheries, value=TRUE, invert=TRUE)

# copy these files over:
file.copy(fisheries, "/home/shares/food-systems/Food_footprint/all_food_systems/analysis/raw")
add_wildcaught <- paste0("/home/shares/food-systems/Food_footprint/all_food_systems/analysis/raw/",basename(fisheries))
file.rename(from= add_wildcaught, 
            to = gsub("/raw/", "/raw/wildcaught_", add_wildcaught))

```

Create feed rasters that describe pressures for each system and feed object:

```{r}

tif_list <- read_csv(here("_analysis/final_list_raw_tiffs.csv")) %>%
  pull(raw_tiffs_list)

feed_list <- grep("_x_", tif_list, value=TRUE) 

feed_list_df <- data.frame(feed_system = feed_list) %>%
  mutate(feed_combo = feed_list) %>%
  separate(feed_system, c("consuming_system", "feed_source"), sep = "_x_") %>%
  mutate(feed_source = gsub(".tif", "", feed_source))

library(foreach)
library(doParallel)

registerDoParallel(6)

foreach(combo=feed_list_df$feed_combo) %dopar% {

#for(combo in feed_list_df$feed_combo){
  #combo = "land_human_human_human_x_marine_fofm_fisheries_meat.tif"
  
  # get the consumer name
  consumer_name <- feed_list_df$consuming_system[feed_list_df$feed_combo == combo]
  
  # get the proportion consumption raster
  prop_tif_file <- list.files('/home/shares/food-systems/Food_footprint/all_food_systems/datalayers', pattern = combo, full=TRUE)
  prop_tif <- raster(prop_tif_file)
  # plot(prop_tif)
  
  # get the pressures for the food source
  food_source_name <- feed_list_df$feed_source[feed_list_df$feed_combo == combo]
  food_source_pressures <- list.files('/home/shares/food-systems/Food_footprint/all_food_systems/datalayers', pattern = food_source_name, full=TRUE)
  food_source_pressures <- grep("_x_", food_source_pressures, invert=TRUE, value=TRUE)
  
  # report if not 4 stressors for each system
  if(length(food_source_pressures)!=4){
    cat(sprintf("incorrect number of pressures for %s, n= %s", food_source_name, length(food_source_pressures)))
  }
  
pressures_stack <- stack(food_source_pressures)
proportion_stressor <- overlay(pressures_stack, prop_tif, fun=function(x,y){x*y})

names(proportion_stressor) <- names(pressures_stack)
# save each stack layer as its own raster
for(i in 1:4) { # i=1
  single_band <- raster(proportion_stressor, layer = i)
  writeRaster(single_band, sprintf("/home/shares/food-systems/Food_footprint/all_food_systems/analysis/raw_feed_combos/feed_%s_x_%s.tif", consumer_name, names(single_band)),
              overwrite=TRUE)
}
}


aux_files <- list.files("/home/shares/food-systems/Food_footprint/all_food_systems/analysis/raw_feed_combos", full=TRUE, pattern = ".aux.xml")
file.remove(aux_files)
```

Collect the human consumption rasters and transfer to the raw file. 
(these are considered complete, the livestock/aquaculture feed will be summed to major category)
```{r}

human <- list.files('/home/shares/food-systems/Food_footprint/all_food_systems/analysis/raw_feed_combos', pattern = "human_human_human", full=TRUE)

# copy these files over, and revise name:
file.copy(human, "/home/shares/food-systems/Food_footprint/all_food_systems/analysis/raw")
add_human <- paste0("/home/shares/food-systems/Food_footprint/all_food_systems/analysis/raw/", basename(human))

add_human_crop <- grep("human_x_marine_fofm", add_human, value=TRUE, invert=TRUE)
add_human_fofm <- grep("human_x_marine_fofm", add_human, value=TRUE)

file.rename(from= add_human_crop, 
            to = gsub("feed_land_human_human_human_x_", "farm_", add_human_crop))

file.rename(from= add_human_fofm, 
            to = gsub("feed_land_human_human_human_x_", "wildcaught_", add_human_fofm))  # this will replace the wildcaught marine fofm with only the human consumption


```


Summarize feed for each animal system, with rasters describing pressures for: fodder, marine fofm, and crops.

Function for following:
```{r}

feed_stressor <- function(system_files, pressure="ghg"){
  system_stressor <- grep(pressure, system_files, value=TRUE)
  
  if(length(grep("fofm", system_stressor))>0){
    fofm_stressor <- grep("fofm", system_stressor, value=TRUE)
    fofm_raster <- raster(fofm_stressor)
    writeRaster(fofm_raster, sprintf("/home/shares/food-systems/Food_footprint/all_food_systems/analysis/raw/feedfofm_%s_%s.tif", system_name, pressure), overwrite=TRUE)
  }
  if(length(grep("fodd", system_stressor))>0){
    fodd_stressor <- grep("fodd", system_stressor, value=TRUE)
    fodd_raster <- raster(fodd_stressor)
    writeRaster(fodd_raster, sprintf("/home/shares/food-systems/Food_footprint/all_food_systems/analysis/raw/feedfodd_%s_%s.tif", system_name, pressure), overwrite=TRUE)
  }
  
  crop_stressor <- grep("fodd|fofm", system_stressor, value=TRUE, invert=TRUE)
  crop_stressor_stack <- stack(crop_stressor)
  check <- raster::calc(crop_stressor_stack, fun=sum, na.rm=TRUE, overwrite=TRUE,
                        filename = sprintf("/home/shares/food-systems/Food_footprint/all_food_systems/analysis/raw/feedcrop_%s_%s.tif", system_name, pressure))
}


```

```{r}
# get the list of animal systems
fed_system_list <- read_csv(here("_analysis/food_system_categories.csv")) %>%
  filter(fed=="fed") %>%
  mutate(system = paste(origin, organism, system, product, sep="_")) %>%
  pull(system)

# feed combos

for(system_name in fed_system_list){
#system_name <- fed_system_list[20]
system_files <- list.files("/home/shares/food-systems/Food_footprint/all_food_systems/analysis/raw_feed_combos", pattern = system_name, full=TRUE)
feed_stressor(system_files, pressure="ghg")
feed_stressor(system_files, pressure="disturbance")
feed_stressor(system_files, pressure="water")
feed_stressor(system_files, pressure="nutrient")
}
          
aux_files <- list.files("/home/shares/food-systems/Food_footprint/all_food_systems/analysis/raw", full=TRUE, pattern = ".aux.xml")
file.remove(aux_files)

```
