==> _analysis/STEP4_rescaling_values.Rmd <==
---
title: "Scaling values"
output: html_document
editor_options: 
  chunk_output_type: console
---

Scaling is based on the global total of each pressure due to food production.

Here I will calculate the total global total of each pressure.  This will be saved as a csv file and used in subsequent scripts to rescale the rasters.

I use the original lat/long coordinate reference system rasters because these values are the least manipulated and should be the most accurate (although this really will not make a difference).

Run time: 20-30min


```{r setup, include=FALSE}

library(rgdal)
library(sp)
library(raster)
library(tidyverse)
library(here)
```

```{r}

pressure_list <- c("ghg", "nutrient", "disturbance", "water")

rescale_data <- data.frame(pressure = pressure_list, global_total = NA)

for(pressure in pressure_list){ # pressure = "water"
food_files_pressure <- list.files("/home/shares/food-systems/Food_footprint/all_food_systems/analysis/raw", full=TRUE, pattern = pressure)

# farm first:
food_files_farm <- grep("farm", food_files_pressure, value=TRUE)
stack_files_farm <- stack(food_files_farm)

raster::calc(stack_files_farm, fun=sum, na.rm=TRUE, 
filename = sprintf("/home/shares/food-systems/Food_footprint/all_food_systems/analysis/tmp/farm_pressure_%s.tif", pressure),
overwrite=TRUE, progress="text")

# food second:
food_files_feed <- grep("feedcrop|feedfodd|feedfofm", food_files_pressure, value=TRUE)
stack_files_feed <- stack(food_files_feed)

raster::calc(stack_files_feed, fun=sum, na.rm=TRUE, 
filename = sprintf("/home/shares/food-systems/Food_footprint/all_food_systems/analysis/tmp/feed_pressure_%s.tif", pressure),
overwrite=TRUE, progress="text")

# wildcaught third:
food_files_fish <- grep("wildcaught", food_files_pressure, value=TRUE)
stack_files_fish <- stack(food_files_fish)

raster::calc(stack_files_fish, fun=sum, na.rm=TRUE, 
filename = sprintf("/home/shares/food-systems/Food_footprint/all_food_systems/analysis/tmp/fish_pressure_%s.tif", pressure),
overwrite=TRUE, progress="text")

## combine farm, feed, fish
total_pressure_stack <- stack(sprintf("/home/shares/food-systems/Food_footprint/all_food_systems/analysis/tmp/fish_pressure_%s.tif", pressure),
      sprintf("/home/shares/food-systems/Food_footprint/all_food_systems/analysis/tmp/feed_pressure_%s.tif", pressure),
      sprintf("/home/shares/food-systems/Food_footprint/all_food_systems/analysis/tmp/farm_pressure_%s.tif", pressure))

total_pressure <- raster::calc(total_pressure_stack, fun=sum, na.rm=TRUE, progress="text")

sum_stressor <- cellStats(total_pressure, stat='sum', na.rm=TRUE)

rescale_data$global_total[rescale_data$pressure == pressure] <- sum_stressor
}

write_csv(rescale_data, here("_analysis/rescale_values.csv"))
```
