==> _analysis/STEP7_summary_df_rasters.Rmd <==
---
title: "Raster extract"
output: html_document
editor_options: 
  chunk_output_type: console
---

extracting data from / home/shares/food-systems/Food_footprint/all_food_systems/analysis/raw

```{r}

library(sf)
library(RColorBrewer)
library(here)

source(here("_workflow/common.R"))
source(here("_workflow/common_spatial.R"))
```


## Summed stressors for each region

Use land/eez raster to extract the summed pressures on land in eez areas.

```{r}
rgns_ezz_land <- raster(file.path(prep, "spatial/land_eez_rgns.tif"))
rgns_ezz_land <-calc(rgns_ezz_land, fun = function(x){ifelse(is.na(x), 9999, x)})

par(mar=c(1,1,1,1))
plot(rgns_ezz_land)

rescaling_values <- read_csv(here("_analysis/rescale_values.csv")) 

food_rgns <- food_rgns %>% 
  add_row(iso3c = "HSX", ID_0 = 9999, Country = "High Seas")
```

Create one big df with zonal stats of all our layers by country

```{r}
summary_df <- data.frame(country = NA,
                         iso3c = NA,
                         category = NA,
                         origin=NA, 
                         organism=NA,
                         system = NA, 
                         product = NA,
                         pressure = NA,
                         sum = NA)

file_list <- list.files(path = file.path("/home/shares/food-systems/Food_footprint/all_food_systems/analysis/raw"), full.names = TRUE)

for(file in file_list){
  
#  file = file_list[1]
  
  country_sum <- zonal(raster(file), rgns_ezz_land, fun="sum", progress="text", na.rm=TRUE) %>% 
    data.frame() %>%
    rename(ID_0 = zone) %>%
    left_join(food_rgns, by="ID_0")
  
    cat <- str_split(str_split(file, pattern = "/")[[1]][9], pattern = "_")[[1]][1]
    ori <- str_split(str_split(file, pattern = "/")[[1]][9], pattern = "_")[[1]][2]
    org <- str_split(str_split(file, pattern = "/")[[1]][9], pattern = "_")[[1]][3]
    sys <- str_split(str_split(file, pattern = "/")[[1]][9], pattern = "_")[[1]][4]
    pro <- str_split(str_split(file, pattern = "/")[[1]][9], pattern = "_")[[1]][5]
    pre <- str_split(str_split(file, pattern = "/")[[1]][9], pattern = "_")[[1]][6]
    pre <- str_split(pre, pattern = ".tif")[[1]][1]
  
country_sum_clean <- country_sum %>% 
  mutate(category = cat,
          origin = ori,
          organism = org,
          system = sys,
          product = pro,
          pressure = pre) %>% 
  select(-ID_0) %>% 
  rename(country = Country) %>% 
  relocate(sum, .after = last_col()) 

  summary_df <- rbind(summary_df, country_sum_clean) %>% 
  filter(!is.na(product))
  
  write_csv(summary_df, file.path("/home/shares/food-systems/Food_footprint/all_food_systems/analysis/raw_df_summary/rgn_raw_summary.csv"))
}

```


Make it so that the "origin" column is where is came from and a "consumed" column where is ends up (land/ocean)
```{r}
summary_df <- read_csv(file.path("/home/shares/food-systems/Food_footprint/all_food_systems/analysis/raw_df_summary/rgn_raw_summary.csv")) %>%
  mutate(origin_raw = origin) %>% 
  rename(consumed = origin) %>% 
  mutate(origin = case_when(category == "farm"  & origin_raw == "marine" ~ "marine",
                            category == "farm"  & origin_raw == "land" ~ "land",
                            category == "feedcrop" ~ "land",
                            category == "feedfodd" ~ "land",
                            category == "feedfofm" ~ "marine",
                            category == "wildcaught" & consumed == "freshwater" ~ "freshwater",
                            category == "wildcaught" & consumed == "marine" ~ "marine")) %>% 
  select(-origin_raw)
  
write_csv(summary_df, file.path("/home/shares/food-systems/Food_footprint/all_food_systems/analysis/raw_df_summary/rgn_raw_summary.csv"))

write_csv(summary_df, here("_analysis/rgn_raw_summary.csv"))
  
```
